// types/employee.ts
export type Employee = {
  employee_id: number;
  department: string;
  region: string;
  education: string;
  gender: string;
  recruitment_channel: string;
  no_of_trainings: number;
  age: number;
  previous_year_rating: number;
  length_of_service: number;
  awards_won: number;
  avg_training_score: number;
  predicted_promotion: boolean;
};